backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
celery==5.3.4
redis==5.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pytrends==4.9.2
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
jinja2==3.1.2
sendgrid==6.10.0
slack-sdk==3.26.1
